A replay is now sent to joining clients. #66
A client can now join a game that was started, and participate normally,
as if it was there from the start.
Next commit should fix the issue of a leaving client, make the game pause
for everyone else, if it left before confirming an action that was sent
to it.